docs: refactored the repo files for the showcase example
