+++ packages/docs/src/examples/showcase/MUIRepos.tsx
@@ -1,15 +1,40 @@
 import React from 'react';
-import {ComponentsRepo, RootSchemaConsumer, isRequired} from '@autoviews/core';
+import {ComponentsRepo, isRequired, RootSchemaConsumer} from '@autoviews/core';
 import {Box} from '@mui/material';
+import {TableCell, Avatar} from '@mui/material';
 
-import {MUIForm, MUIText, MUINumber, MUISwitch, MUISlider} from './MUIForm';
-import {
-    BootstrapForm,
-    BootstrapText,
-    BootstrapNumber,
-    BootstrapSwitch
-} from './BootstrapForm';
-import {detectEnums} from './basicRepo';
+import {MUITable, MUITableRow} from './MUITable';
+import {IMAGE_SUBTYPE} from './schemas';
+import {basicRepo, detectEnums} from './basicRepo';
+import {MUIForm, MUINumber, MUISlider, MUISwitch, MUIText} from './MUIForm';
+
+export const MUITableRepo = basicRepo
+    .clone('MUITableRepo')
+    .register('array', {
+        name: 'tableComponent',
+        component: MUITable,
+        predicate: node => node.items.type === 'object'
+    })
+    .register('object', {
+        name: 'tableRowComponent',
+        component: MUITableRow
+    })
+    .register('string', {
+        name: 'avatarComponent',
+        component: props => <Avatar src={props.data} />,
+        predicate: node => node.format === IMAGE_SUBTYPE
+    })
+    .addWrapper(item => <TableCell>{item}</TableCell>, {
+        include: [
+            'textComponent',
+            'numberComponent',
+            'booleanComponent',
+            'imageComponent',
+            'emailComponent',
+            'linkComponent',
+            'avatarComponent'
+        ]
+    });
 
 export const MUIFormRepo = new ComponentsRepo('MUIFormRepo', detectEnums)
     .register('object', {
@@ -69,35 +94,3 @@ export const MUIFormRepo = new ComponentsRepo('MUIFormRepo', detectEnums)
             ]
         }
     );
-
-export const BootstrapFormRepo = new ComponentsRepo(
-    'BootstrapFormRepo',
-    detectEnums
-)
-    .register('object', {
-        name: 'formComponent',
-        component: BootstrapForm
-    })
-    .register('string', {
-        name: 'textComponent',
-        component: BootstrapText
-    })
-    .register('number', {
-        name: 'numberComponent',
-        component: BootstrapNumber
-    })
-    .register('boolean', {
-        name: 'switchComponent',
-        component: BootstrapSwitch
-    })
-    .register('array', {
-        name: 'arrayComponent',
-        component: () => <span>array</span>
-    })
-    .register('oneOfEnumLike', {
-        name: 'enumComponent',
-        component: props => <span>{props.data}</span>
-    })
-    .addWrapper(item => <div style={{margin: '5px 0'}}>{item}</div>, {
-        include: ['textComponent', 'numberComponent', 'switchComponent']
-    });
